+++ lua/obsidian/lsp/init.lua
@@ -10,7 +10,7 @@ lsp.start = function(buf)
   local capabilities = vim.lsp.protocol.make_client_capabilities()
   capabilities.workspace = capabilities.workspace or {}
   capabilities.workspace.fileOperations =
-    vim.tbl_extend("force", capabilities.workspace.fileOperations or {}, { didRename = true })
+    vim.tbl_extend("force", capabilities.workspace.fileOperations or {}, { didRename = true, willDelete = true })
   -- manually enable dynamic registration for file watching, since neovim turns off this capability by default on linux and BSD
   capabilities.workspace.didChangeWatchedFiles = {
     dynamicRegistration = true,